requirements.txt
fastapi
uvicorn
pydantic
requests
pandas
python-dotenv
gspread
oauth2client
slack_sdk
vaderSentiment
textstat
langdetect
language_tool_python
scikit-learn
textblob
transformers
torch
imblearn
groq
genai
streamlit